WIP refactor state into application
